kika-api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
matplotlib>=3.8.0
numpy>=1.26.0
scipy>=1.11.0
kika-nd
pyinstaller
